move catalog constants to skylet constants
